Changed from int to small int field capacity in classroom

+++ alembic/alembic/versions/32edd08dd080_create_tables.py
@@ -36,7 +36,7 @@ def upgrade() -> None:
         sa.Column('id', sa.Integer, primary_key=True),
         sa.Column('name', sa.VARCHAR(32)),
         sa.Column('classroomStatus', sa.Enum('available', 'unavailable'), default='available'),
-        sa.Column('capacity', sa.Integer)
+        sa.Column('capacity', sa.SmallInteger)
     )
 
     op.create_table(
