+++ classroom_booking/models.py
@@ -2,7 +2,7 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session
-from sqlalchemy import Column, Integer, String, Date, ForeignKey, Enum, DateTime
+from sqlalchemy import Column, Integer, String, Date, ForeignKey, Enum, DateTime, SmallInteger
 
 DB_URL = "mysql://root:E100_amx1390_maus@localhost:3306/ap"
 
@@ -31,7 +31,7 @@ class Classroom(BaseModel):
 	id = Column(Integer, primary_key=True)
 	name = Column(String(32))
 	classroomStatus = Column(Enum('available', 'unavailable'), default='available')
-	capacity = Column(Integer)
+	capacity = Column(SmallInteger)
 
 
 class Order(BaseModel):
